# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the authenticated user's notifications, newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "body": {
                                        "type": "string"
                                    },
                                    "created_at": {
                                        "type": "string"
                                    },
                                    "data": {
                                        "type": "string"
                                    },
                                    "id": {
                                        "type": "string"
                                    },
                                    "read": {
                                        "type": "boolean"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notifications/{id}/read": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks a notification as read. Validates ownership against the JWT subject.",
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark notification read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/threads": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated WhatsApp conversation threads with contact details.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp"
                ],
                "summary": "List WhatsApp threads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by status: open|pending|closed",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.WhatsAppThread"
                            }
                        }
                    }
                }
            }
        },
        "/threads/{id}/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets thread_status to closed.",
                "tags": [
                    "WhatsApp"
                ],
                "summary": "Close thread",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/threads/{id}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns messages for a WhatsApp thread, ordered by sent_at ascending.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WhatsApp"
                ],
                "summary": "Get thread messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Thread UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max messages to return (default 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.WhatsAppMessage"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_dynamicweblab_masaar-crm_internal_domain.Contact": {
            "type": "object",
            "properties": {
                "assigned_to": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "description": "\"ar\" | \"en\"",
                    "type": "string"
                },
                "lead_score": {
                    "type": "integer"
                },
                "phone_wa": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.MessageDirection": {
            "type": "string",
            "enum": [
                "inbound",
                "outbound"
            ],
            "x-enum-varnames": [
                "DirectionInbound",
                "DirectionOutbound"
            ]
        },
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.ThreadStatus": {
            "type": "string",
            "enum": [
                "open",
                "closed",
                "pending"
            ],
            "x-enum-varnames": [
                "ThreadOpen",
                "ThreadClosed",
                "ThreadPending"
            ]
        },
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.WhatsAppMessage": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "direction": {
                    "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.MessageDirection"
                },
                "id": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "thread_id": {
                    "type": "string"
                },
                "wa_message_id": {
                    "type": "string"
                }
            }
        },
        "github_com_dynamicweblab_masaar-crm_internal_domain.WhatsAppThread": {
            "type": "object",
            "properties": {
                "ai_summary": {
                    "type": "string"
                },
                "contact": {
                    "description": "Joined",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.Contact"
                        }
                    ]
                },
                "contact_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_message_at": {
                    "type": "string"
                },
                "message_count": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.WhatsAppMessage"
                    }
                },
                "thread_status": {
                    "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.ThreadStatus"
                },
                "wa_account_id": {
                    "type": "string"
                }
            }
        }
    }
}